requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu/
Django~=4.2
django-cors-headers~=4.3
djangorestframework~=3.14
#medcat~=1.10
medcat@git+https://github.com/CogStack/MedCAT@9ea067549ff41242aaf4df53236221a0f2c06898
pydantic~=1.10

